requirements.txt
# Python packages and version used in development environment

alabaster==0.7.8
arrow==0.8.0
attrs==16.0.0
Babel==2.3.3
circus==0.13.0
coverage==4.0.3
docutils==0.12
iowait==0.2
Jinja2==2.8
MarkupSafe==0.23
psutil==4.2.0
py==1.4.31
Pygments==2.1.3
pytest==2.8.1
PyYAML==3.11
pyzmq==15.2.0
requests==2.11.1
schedule==0.3.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
tornado==4.3
